requirements.txt
openai>=1.0.0
pyyaml>=6.0
langgraph>=0.0.15
typing_extensions>=4.5.0